Move remaining positioning logic local to hold note blueprint
